+++ core/SEOstats/Common/SEOstatsException.php
@@ -1,3 +1,0 @@
-<?php
-namespace SEOstats\Common;
-class SEOstatsException extends \Exception{}